+++ liga-futeba/src/main/java/br/com/futeba/models/Place.java
@@ -10,81 +10,81 @@ import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
 
 @Entity(name = "place")
-public class Place implements Serializable{
+public class Place implements Serializable {
 
 	private static final long serialVersionUID = 2062908763035875940L;
 
 	@Id
-    @Column(name = "place_id", nullable = false)
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+	@Column(name = "place_id", nullable = false)
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private long id;
 
-    @NotNull
-    private String name;
+	@NotNull
+	private String name;
 
-    private String type;
-    private String address;
-    private String city;
-    private String neighborhood;
-    private String zipCode;
+	private String type;
+	private String address;
+	private String city;
+	private String neighborhood;
+	private String zipCode;
 
-    public Place() {
-    }
+	public Place() {
+	}
 
-    public Integer getId() {
-        return id;
-    }
+	public long getId() {
+		return id;
+	}
 
-    public void setId(final Integer id) {
-        this.id = id;
-    }
+	public void setId(final long id) {
+		this.id = id;
+	}
 
-    public String getName() {
-        return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public void setName(final String name) {
-        this.name = name;
-    }
+	public void setName(final String name) {
+		this.name = name;
+	}
 
-    public String getType() {
-        return type;
-    }
+	public String getType() {
+		return type;
+	}
 
-    public void setType(final String type) {
-        this.type = type;
-    }
+	public void setType(final String type) {
+		this.type = type;
+	}
 
-    public String getAddress() {
-        return address;
-    }
+	public String getAddress() {
+		return address;
+	}
 
-    public void setAddress(final String address) {
-        this.address = address;
-    }
+	public void setAddress(final String address) {
+		this.address = address;
+	}
 
-    public String getCity() {
-        return city;
-    }
+	public String getCity() {
+		return city;
+	}
 
-    public void setCity(final String city) {
-        this.city = city;
-    }
+	public void setCity(final String city) {
+		this.city = city;
+	}
 
-    public String getNeighborhood() {
-        return neighborhood;
-    }
+	public String getNeighborhood() {
+		return neighborhood;
+	}
 
-    public void setNeighborhood(final String neighborhood) {
-        this.neighborhood = neighborhood;
-    }
+	public void setNeighborhood(final String neighborhood) {
+		this.neighborhood = neighborhood;
+	}
 
-    public String getZipCode() {
-        return zipCode;
-    }
+	public String getZipCode() {
+		return zipCode;
+	}
 
-    public void setZipCode(final String zipCode) {
-        this.zipCode = zipCode;
-    }
+	public void setZipCode(final String zipCode) {
+		this.zipCode = zipCode;
+	}
 
 }
