+++ liga-futeba/src/main/java/br/com/futeba/repositorys/GameRepository.java
@@ -9,13 +9,13 @@ import org.springframework.data.repository.query.Param;
 
 import br.com.futeba.models.Game;
 
-public interface GameRepository extends JpaRepository<Game, Integer> {
+public interface GameRepository extends JpaRepository<Game, Long> {
 
-    public Optional<Game> findById(@Param("idParam") Integer id);
+	public Optional<Game> findById(@Param("idParam") long id);
 
-    public Optional<Game> saveAndFlush(Optional<Game> game);
-    
-    @Query(value = "SELECT * FROM jogos WHERE YEAR(data) = :year ORDER BY data ASC", nativeQuery = true)
-    public List<Game> getGames(@Param("year") Integer year);
+	public Optional<Game> saveAndFlush(Optional<Game> game);
+
+	@Query(value = "SELECT * FROM jogos WHERE YEAR(data) = :year ORDER BY data ASC", nativeQuery = true)
+	public List<Game> getGames(@Param("year") Integer year);
 
 }
