+++ liga-futeba/src/main/java/br/com/futeba/repositorys/PositionRepository.java
@@ -4,10 +4,9 @@ import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-import br.com.futeba.models.Player;
 import br.com.futeba.models.Position;
 
-public interface PositionRepository extends CrudRepository<Position, Integer> {
+public interface PositionRepository extends CrudRepository<Position, Long> {
 
 	Optional<Position> saveAndFlush(Optional<Position> posicao);
