+++ liga-futeba/src/main/java/br/com/futeba/repositorys/TeamRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.repository.query.Param;
 
 import br.com.futeba.models.Team;
 
-public interface TeamRepository extends JpaRepository<Team, Integer> {
-	
-    @Query(value = "SELECT quadro, SUM(pontos) as pts, SUM(total_gols_mandante) as gm, SUM(total_gols_visitante) as gs, SUM(total_gols_mandante)  - SUM(total_gols_visitante) as sg, COUNT(*) as jgs,  (SUM(pontos) / (COUNT(*) * 3)) as ap FROM jogos WHERE YEAR(data) = :year GROUP BY quadro ORDER BY quadro ASC", nativeQuery = true)
-    public List<?> getTeamStats(@Param("year") Integer year);
+public interface TeamRepository extends JpaRepository<Team, Long> {
+
+	@Query(value = "SELECT quadro, SUM(pontos) as pts, SUM(total_gols_mandante) as gm, SUM(total_gols_visitante) as gs, SUM(total_gols_mandante)  - SUM(total_gols_visitante) as sg, COUNT(*) as jgs,  (SUM(pontos) / (COUNT(*) * 3)) as ap FROM jogos WHERE YEAR(data) = :year GROUP BY quadro ORDER BY quadro ASC", nativeQuery = true)
+	public List<?> getTeamStats(@Param("year") Integer year);
 
 	public Optional<Team> saveAndFlush(Optional<Team> team);
